move unit fixtures and loader to unit test root
