+++ tests/unit/conftest.py
@@ -6,6 +6,8 @@ from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
 import sys
+import os
+import json
 import pytest
 
 
@@ -13,3 +15,27 @@ import pytest
 def skip_python():
     if sys.version_info < (3, 6):
         pytest.skip('Skipping on Python %s. community.hashi_vault supports Python 3.6 and higher.' % sys.version)
+
+
+@pytest.fixture
+def fixture_loader():
+    def _loader(name, parse='json'):
+        here = os.path.dirname(os.path.realpath(__file__))
+        fixture = os.path.join(here, 'fixtures', name)
+
+        if parse == 'path':
+            return fixture
+
+        with open(fixture, 'r') as f:
+            if parse == 'json':
+                d = json.load(f)
+            elif parse == 'lines':
+                d = f.readlines()
+            elif parse == 'raw':
+                d = f.read()
+            else:
+                raise ValueError("Unknown value '%s' for parse" % parse)
+
+        return d
+
+    return _loader
