+++ tests/unit/fixtures/lookup-self_with_meta.json
