+++ tests/unit/fixtures/lookup-self_without_meta.json
