+++ tests/unit/plugins/module_utils/authentication/conftest.py
@@ -5,8 +5,6 @@
 from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
-import os
-import json
 import pytest
 
 try:
@@ -57,27 +55,3 @@ def client():
 @pytest.fixture
 def warner():
     return mock.MagicMock()
-
-
-@pytest.fixture
-def fixture_loader():
-    def _loader(name, parse='json'):
-        here = os.path.dirname(os.path.realpath(__file__))
-        fixture = os.path.join(here, 'fixtures', name)
-
-        if parse == 'path':
-            return fixture
-
-        with open(fixture, 'r') as f:
-            if parse == 'json':
-                d = json.load(f)
-            elif parse == 'lines':
-                d = f.readlines()
-            elif parse == 'raw':
-                d = f.read()
-            else:
-                raise ValueError("Unknown value '%s' for parse" % parse)
-
-        return d
-
-    return _loader
